# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API Trothalo
  description: API hệ thống đặt lưu trú
  version: 1.0.0
  contact:
    name: Trothalo Team
    email: support@trothalo.com
servers:
  - url: http://localhost:8080
    description: Local server
  - url: https://backend.trothalo.click
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
          example: REQUIRED_FIELD
        message:
          type: string
          example: Email không được để trống
        data:
          type: object
          nullable: true
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum: [admin, user, host]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Bank:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        account_number:
          type: string
        account_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Rate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        room_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Holiday:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Đăng ký tài khoản mới
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - phone
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                phone:
                  type: string
      responses:
        "200":
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login:
    post:
      summary: Đăng nhập
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      user:
                        $ref: "#/components/schemas/User"
        "401":
          description: Email hoặc mật khẩu không đúng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Room endpoints
...
  /api/v1/banks:
    get:
      summary: Lấy danh sách ngân hàng
      tags:
        - Banks
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lấy danh sách ngân hàng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bank"
        "401":
          description: Không có quyền truy cập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Thêm ngân hàng mới
      tags:
        - Banks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - account_number
                - account_name
              properties:
                name:
                  type: string
                account_number:
                  type: string
                account_name:
                  type: string
      responses:
        "200":
          description: Thêm ngân hàng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/Bank"
        "401":
          description: Không có quyền truy cập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Withdrawal endpoints
...
  /api/v1/rates:
    get:
      summary: Lấy danh sách đánh giá
      tags:
        - Rates
      parameters:
        - in: query
          name: room_id
          schema:
            type: string
            format: uuid
          description: ID của phòng
      responses:
        "200":
          description: Lấy danh sách đánh giá thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Rate"
        "404":
          description: Không tìm thấy phòng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Tạo đánh giá mới
      tags:
        - Rates
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - room_id
                - rating
                - comment
              properties:
                room_id:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        "200":
          description: Tạo đánh giá thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/Rate"
        "401":
          description: Không có quyền truy cập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Không tìm thấy phòng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Revenue endpoints
...
  /api/v1/holidays:
    get:
      summary: Lấy danh sách ngày nghỉ
      tags:
        - Holidays
      responses:
        "200":
          description: Lấy danh sách ngày nghỉ thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Holiday"

    post:
      summary: Thêm ngày nghỉ mới
      tags:
        - Holidays
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - date
              properties:
                name:
                  type: string
                date:
                  type: string
                  format: date
      responses:
        "200":
          description: Thêm ngày nghỉ thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/Holiday"
        "401":
          description: Không có quyền truy cập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Invoice endpoints
  /api/v1/invoices:
    get:
      summary: Lấy danh sách hóa đơn
      tags:
        - Invoices
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: order_id
          schema:
            type: string
            format: uuid
          description: ID của đơn đặt phòng
      responses:
        "200":
          description: Lấy danh sách hóa đơn thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
        "401":
          description: Không có quyền truy cập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Accommodation endpoints
...